backend-flask/api/requirements.txt
aiohttp==3.7.4.post0
aniso8601==9.0.1
async-timeout==3.0.1
attrdict==2.0.1
attrs==20.3.0
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
greenlet==1.1.0
gunicorn==20.1.0
hexbytes==0.2.1
idna==2.10
inflection==0.5.1
ipfshttpclient==0.7.0a1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==1.1.1
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
parsimonious==0.8.1
phonenumbers==8.12.33
protobuf==3.15.8
pycryptodome==3.10.1
PyJWT==2.1.0
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rlp==2.0.1
rq==1.8.1
six==1.15.0
SQLAlchemy==1.4.17
swagger-ui==0.1.2
swagger-ui-bundle==0.0.8
tinydb==4.4.0
toolz==0.11.1
typing-extensions==3.10.0.0
urllib3==1.26.4
varint==1.0.2
web3==5.19.0
websockets==8.1
Werkzeug==1.0.1
yarl==1.6.3
